+++ tools/language-server/wstest/src/format.rs
@@ -1,7 +1,21 @@
+// === Non-Standard Linter Configuration ===
+#![warn(unsafe_code)]
+#![warn(missing_docs)]
+#![warn(trivial_casts)]
+#![warn(trivial_numeric_casts)]
+#![warn(unused_import_braces)]
+#![warn(unused_qualifications)]
+#![warn(missing_copy_implementations)]
+#![warn(missing_debug_implementations)]
+
 use time::format_description::well_known::Rfc3339;
 use time::OffsetDateTime;
 
 
+
+// =================
+// === Constants ===
+// =================
 
 pub static MESSAGE_BINARY: &str = "<binary>";
 
@@ -13,6 +27,12 @@ static RESPONSE_IGNORED: &str = "wstest ignored response";
 
 static FMT_LEVEL: &str = "info";
 static FMT_MODULE: &str = "main";
+
+
+
+// ==========================
+// === Message Formatting ===
+// ==========================
 
 /// Message for logging the initialization request
 pub fn init_request(message: &str) -> String {